+++ src/pages/addCard.jsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { Link } from "react-router-dom";
-import Title from "../stories/Title";
-import Button from "../stories/Button";
-import Input from "../stories/Input";
+import Title from "../components/Title";
+import Button from "../components/Button";
+import Input from "../components/Input";
 
 function AddCard() {
   const [cardNumberSeries, setCardNumberSeries] = useState("");
