+++ src/stories/Button.stories.jsx
@@ -1,4 +1,4 @@
-import Button from "./Button";
+import Button from "../components/Button";
 
 export default {
   title: "Components/Button",
