+++ src/stories/Input.stories.jsx
@@ -1,4 +1,4 @@
-import Input from "./Input";
+import Input from "../components/Input";
 
 export default {
   title: "Components/Input",
